requirements.txt
transformers==4.46.1
imagecodecs==2024.9.22
torchgeo==0.7.0.dev0
scikit-learn== 1.5.2
tifffile==2024.9.20
seaborn==0.13.2
albumentations==1.4.21
timm==0.9.7
matplotlib==3.9.2
tqdm==4.66.6      
pandas==2.2.3
opencv-python==4.10.0
Pillow==11.0.0
requests==2.32.3
